FRONT: move Home page to Services page
- remove bg color for debug
- add area zone to services page

MINOR
